途中 #212

+++ system/core/myfirestore/firestore_controller.go
@@ -53,7 +53,12 @@ func (c *FirestoreController) update(ctx context.Context, tx *firestore.Transact
 	}
 }
 
-func (c *FirestoreController) delete(ctx context.Context, tx *firestore.Transaction, ref *firestore.DocumentRef, opts ...firestore.Precondition) error {
+func (c *FirestoreController) delete(ctx context.Context, tx *firestore.Transaction, ref *firestore.DocumentRef, refExists bool, opts ...firestore.Precondition) error {
+	// ドキュメントが元から存在しないときは明示的にエラーを返す
+	if !refExists {
+		return errors.New("ref have to exists before delete.")
+	}
+	
 	if tx != nil {
 		return tx.Delete(ref, opts...)
 	} else {
@@ -371,9 +376,9 @@ func (c *FirestoreController) UpdateSeat(tx *firestore.Transaction, seat SeatDoc
 	return c.set(nil, tx, ref, seat)
 }
 
-func (c *FirestoreController) RemoveSeat(tx *firestore.Transaction, seatId int) error {
+func (c *FirestoreController) RemoveSeat(tx *firestore.Transaction, seatId int, seatExists bool) error {
 	ref := c.seatsCollection().Doc(strconv.Itoa(seatId))
-	return c.delete(nil, tx, ref)
+	return c.delete(nil, tx, ref, seatExists)
 }
 
 func (c *FirestoreController) AddLiveChatHistoryDoc(ctx context.Context, tx *firestore.Transaction,
